Streamline Run initialization
Whenever we read an input file, we hang on to a copy of the initial
tree.  Then, both the initial and subsequent instantiations of the run
start from this initial tree (so we no longer have to hang on to the
input file contents, or reparse them on option changes; additionally,
runs loaded from a .dphy file can now be restarted with different
options).  All code paths go through instantiateRun(), which is the one
place where a Run is created and runParameters are applied.  The actual
run parameters are centralized in Pythia.runParameters and are frozen
for the length of a run.  Any changes in the RunnerUI work by creating a
modified copy of the current runParameters and instantiating a brand new
run with it.

TO FIX:
- There is some issue in qcpanel.ts when resetting a run, where setQCDiv
obliterates a template that was never cloned, so setQCDiv cannot be
called twice (?)